Create new `runty8-event-loop2` crate (temp name)
