+++ src/runty8-event-loop2/Cargo.toml
@@ -1,23 +1,20 @@
 [package]
-name = "runty8-event-loop"
+name = "runty8-event-loop2"
 version = "0.1.0"
 edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 runty8-core = { path = "../runty8-core" }
 runty8-winit = { path = "../runty8-winit" }
 winit = "0.26"
-instant = { version = "0.1", features = ["wasm-bindgen"] }
-# wasm related stuff
 glow = "0.11"
-wasm-bindgen = { version = "0.2" }
+instant = { version = "0.1", features = ["wasm-bindgen"] }
+# For now
+glium = "0.31"
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+wasm-bindgen = {version = "0.2" }
 web-sys = { version = "0.3", features = ["HtmlCanvasElement", "WebGl2RenderingContext", "Window", "console", "Document"] }
 log = "0.4"
 console_log = "0.2"
 console_error_panic_hook = "0.1.7"
-
-#native only
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-glutin = "0.28.0"
